fix public asset warning by converting public assets to src assets
The warning:
Assets in public directory cannot be imported from JavaScript.
If you intend to import that asset, put the file in the src directory, and use /src/vite.svg instead of /public/vite.svg.
If you intend to use the URL of that asset, use /vite.svg?url.
Files in the public directory are served at the root path.
Instead of /public/vite.svg, use /vite.svg.

Docs:
https://vitejs.dev/guide/assets.html
https://vitejs.dev/guide/features#static-assets